+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [[0.4.0] 2017-09-27](https://github.com/arturock/vscode-gitstash/tree/v0.4.0)
+## [[0.5.0] 2017-10-01](https://github.com/arturock/vscode-gitstash/tree/v0.5.0)
+ - Add missing commands
+   - Save simple, keep index, include untracked, stash all
+   - Pop w/ reindex
+   - Apply w/ reindex
+   - Branch
+- Explorer Tree
+  - Show stashed untracked files too
+  - Update icons
+
+## [[0.4.0] 2017-09-29](https://github.com/arturock/vscode-gitstash/tree/v0.4.0)
 - Add commands
   - Apply
   - Drop